refactor: 노드 상세보기 레이아웃 페이지와 사이드바에서 공용으로 사용하도록 변경
