+++ frontend/src/app/nodes/[id]/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation';
+
+export default function NodePage({ params }: { params: { id: string } }) {
+  redirect(`/nodes/${params.id}/note`);
+}
