+++ frontend/src/components/node-datail/NodeMeetingTab.tsx
@@ -1,0 +1,5 @@
+export const NodeMeetingTab = () => {
+  return <>임시 회의 페이지</>;
+};
+
+export default NodeMeetingTab;
