+++ frontend/src/components/node-datail/NodeNoteTab.tsx
@@ -1,0 +1,5 @@
+export const NodeNoteTab = () => {
+  return <>임시 노트 페이지</>;
+};
+
+export default NodeNoteTab;
